requirements.txt
faker
fastapi
uvicorn
fastapi_pagination
pytest
flake8
black
isort
pandas
requests
apache-airflow
pytest-cov